+++ StrainLink.hpp
@@ -19,9 +19,9 @@ class StrainLink
         );
 
         void update();
-        float update_position();
+        double update_position();
         void update_orientation();
-        float getStrain();
+        double getStrain();
         void resetLambda();
         b2Vec2 getVector();
         std::pair<int,int> getId();
@@ -32,10 +32,10 @@ class StrainLink
         std::shared_ptr<particle> particle_a;
         std::shared_ptr<particle> particle_b;
         std::pair<int, int> id;
-        float compliance;
-        float normalized_compliance;
-        float rest_length;
-        float lambda;
+        double compliance;
+        double normalized_compliance;
+        double rest_length;
+        double lambda;
 };
 
 inline StrainLink::StrainLink() {}
@@ -44,7 +44,7 @@ inline void StrainLink::resetLambda() {
     this->lambda = 0.;
 }
 
-inline float StrainLink::getStrain() {
+inline double StrainLink::getStrain() {
     return log(this->getVector().Length() / this->rest_length);
 }
