+++ params.cpp
@@ -3,21 +3,21 @@
 #include <cassert>
 
 namespace params{
-    float time_step = 1.;
-    float inv_time_step = 1.;
-    float strain_compliance = 1.;
-    float normalized_strain_compliance = 1.;
+    double time_step = 1.;
+    double inv_time_step = 1.;
+    double strain_compliance = 1.;
+    double normalized_strain_compliance = 1.;
     int constraint_iters = 1;
 
 
-    void setTimeStep(float in_time_step) {
+    void setTimeStep(double in_time_step) {
         assert(in_time_step > 0.);
         params::time_step = in_time_step;
         params::inv_time_step = 1./in_time_step;
         void updateNormalizedCompliances();
     }
 
-    void setStrainCompliance(float in_strain_compliance) {
+    void setStrainCompliance(double in_strain_compliance) {
         params::strain_compliance = in_strain_compliance;
         void updateNormalizedCompliances();
     }
