+++ params.hpp
@@ -2,14 +2,14 @@
 // #define PARAMS_HPP
 
 namespace params {
-    extern float time_step;
-    extern float inv_time_step;
-    extern float strain_compliance;
-    extern float normalized_strain_compliance;
+    extern double time_step;
+    extern double inv_time_step;
+    extern double strain_compliance;
+    extern double normalized_strain_compliance;
     extern int constraint_iters;
 
-    void setTimeStep(float in_time_step);
-    void setStrainCompliance(float in_strain_compliance);
+    void setTimeStep(double in_time_step);
+    void setStrainCompliance(double in_strain_compliance);
     void updateNormalizedCompliances();
     void setConstraintIters(int in_constraint_iters);
 }
